third_party/blink/web_tests/virtual/controls-refresh/calendar-picker/date-picker-month-switcher.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="date" id="date-0" value="2019-02-14">
<input type="date" id="date-1" value="2019-02-14">
<script>

promise_test(() => {
  let dateElement = document.getElementById('date-0');
  return openPickerWithPromise(dateElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickMonthPopupButton();
    eventSender.keyDown('ArrowRight');
    eventSender.keyDown('ArrowDown');
    eventSender.keyDown('Enter');
    clickDayCellAt(2, 3);

    assert_equals(dateElement.value, "2019-07-23", "Month chooser button should have changed month");
    assert_equals(internals.pagePopupWindow, null, "Clicking a date should dismiss popup.");
  });
}, "Date picker: Month chooser should allow user to chooser month");

promise_test(() => {
  let dateElement = document.getElementById('date-1');
  return openPickerWithPromise(dateElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickMonthPopupButton();
    eventSender.keyDown('ArrowRight');
    eventSender.keyDown('ArrowDown');
    eventSender.keyDown('Escape');
    clickDayCellAt(2, 3);

    assert_equals(dateElement.value, "2019-02-19", "Escape key should have dismissed month chooser without changing month ");
  });
}, "Date picker: Month chooser should dismiss on Escape key");

</script>
</body>
</html>
